Milestone 46: a heap that stays fast, and one kstring for both arches
kheap64.c is a rewrite rather than a port, because the thing worth
porting was the bug. kheap.c keeps one list of every block, used and
free, and kmalloc walks it from the start: N allocations cost N squared.
Milestone 43 measured that as Wine's syscall rate collapsing 21x during
startup. Carrying it to the architecture meant to run a browser was not
defensible.

Two structures instead: blocks doubly linked in address order, so kfree
coalesces with its neighbours in O(1) instead of rescanning the heap;
and a free list, which is the only thing kmalloc walks. Free-list
pointers live in the payload of the free block, so they cost nothing
while a block is in use - the only reason the minimum allocation is 16
bytes.

The test prints the measurement rather than the claim: 1000 steps for
1000 allocations, where quadratic would be ~500000. And after a thousand
alloc/free pairs plus a 128KB allocation that grew the heap, the heap
collapses to a single free block with nothing used - no fragmentation
left. Coalescing is checked from outside, by requiring the free-list
length to return to what it was; with the merges removed it fails and
the count goes 1 -> 1006.

No console in the 64-bit tree, so exhaustion returns NULL and shows in
the counters. The 32-bit version prints, which is part of what made it
non-portable.

kstring.c is now shared by both builds rather than forked. It needed one
change: size parameters are size_t, not uint32_t. Same type on i386, so
the 32-bit build is unaffected; on x86-64 a uint32_t length would have
capped every copy in the kernel at 4GB. kdiv64 turned out not to be
i386-only - it wraps a 32-bit divl, legal in long mode - so it is
compiled and checked rather than excluded.

Verified on the 32-bit side, since kstring is used tree-wide: clean make
with zero warnings under -Wall -Wextra, make test passes (64 fs + 52 wm
checks), make test-qemu passes all 17 transcript assertions.

64-bit boot test: 79 assertions, 0 failures, up from 56.

+++ include/kstring.h
@@ -15,27 +15,31 @@
  * synthesize calls to those four names, and having our own definitions
  * collide with the compiler's assumptions about them is a subtle trap. */
 
-void  kmemcpy(void* dst, const void* src, uint32_t n);
-void  kmemmove(void* dst, const void* src, uint32_t n);
-void  kmemset(void* dst, uint8_t value, uint32_t n);
-int   kmemcmp(const void* a, const void* b, uint32_t n);
-void* kmemchr(const void* s, uint8_t c, uint32_t n);
+/* Sizes are size_t, not uint32_t, so that one implementation serves both
+ * architectures. On i386 the two are the same type, so nothing about the
+ * 32-bit build changes; on x86-64 a uint32_t length would silently cap
+ * every copy in the kernel at 4GB. */
+void  kmemcpy(void* dst, const void* src, size_t n);
+void  kmemmove(void* dst, const void* src, size_t n);
+void  kmemset(void* dst, uint8_t value, size_t n);
+int   kmemcmp(const void* a, const void* b, size_t n);
+void* kmemchr(const void* s, uint8_t c, size_t n);
 
-uint32_t kstrlen(const char* s);
-uint32_t kstrnlen(const char* s, uint32_t max);
+size_t kstrlen(const char* s);
+size_t kstrnlen(const char* s, size_t max);
 int   kstrcmp(const char* a, const char* b);
-int   kstrncmp(const char* a, const char* b, uint32_t n);
+int   kstrncmp(const char* a, const char* b, size_t n);
 /* Case-insensitive, ASCII only. Milestone 39: HTTP header names are
  * case-insensitive by specification, and a server that answers
  * "content-length" in lower case is not a server that is wrong. */
-int   kstrncasecmp(const char* a, const char* b, uint32_t n);
+int   kstrncasecmp(const char* a, const char* b, size_t n);
 /* ASCII case-insensitive - Win32 module and export lookups are
  * case-insensitive, and import tables spell DLL names every which way
  * ("KERNEL32.dll", "kernel32.DLL", ...). */
 int   kstricmp(const char* a, const char* b);
 char* kstrcpy(char* dst, const char* src);
 /* Bounded copy that always NUL-terminates (unlike strncpy). Returns dst. */
-char* kstrlcpy(char* dst, const char* src, uint32_t size);
+char* kstrlcpy(char* dst, const char* src, size_t size);
 char* kstrcat(char* dst, const char* src);
 char* kstrchr(const char* s, char c);
 char* kstrrchr(const char* s, char c);
